pyproject.toml
[tool.poetry]
name = "hiring"
version = "1.0.0"
description = "Fix this project and get a job at melli.com!"
authors = ["Melli <melli@melli.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = { extras = ["standard"], version = "^0.16.0" }
pydantic = "^1.8.2"
python-multipart = "^0.0.5"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
tomlkit = "^0.7.2"
spacy = "^3.0.6"
en_core_web_sm = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-depends = "^1.0.1"
httpx = "^0.23.0"
pylint = "^2.12.2"
black = "^21.12b0"
isort = "^5.10.1"

[tool.pytest.ini_options]
markers = ["bonus: optional task applicants don't have to solve"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
